+++ components/index.js
@@ -52,6 +52,7 @@ export * from './tbt-address.js';
 /* Display */
 export * from './tbt-approval-flow.js';
 export * from './tbt-audit-log.js';
+export * from './tbt-timeline.js';
 export * from './tbt-field.js';
 export * from './tbt-field-grid.js';
 export * from './tbt-section.js';